modularizing code
