+++ isogeo.py
@@ -31,11 +31,10 @@ from logging.handlers import RotatingFileHandler
 import platform  # about operating systems
 from collections import OrderedDict
 from functools import partial
-import db_manager.db_plugins.postgis.connector as con
 import operator
 
 # PyQT
-# from de QByteArray
+# from QByteArray
 from PyQt4.QtCore import QSettings, QTranslator, qVersion, QCoreApplication, \
     Qt, QByteArray, QUrl
 # from QMessageBox
@@ -45,6 +44,7 @@ from PyQt4.QtGui import QAction, QIcon, QMessageBox, QTableWidgetItem, \
 from PyQt4.QtNetwork import QNetworkRequest
 
 # PyQGIS
+import db_manager.db_plugins.postgis.connector as con
 from qgis.utils import iface, plugin_times, QGis, reloadPlugin
 from qgis.core import QgsNetworkAccessManager, QgsPoint, \
     QgsCoordinateReferenceSystem, QgsCoordinateTransform, QgsVectorLayer, \
@@ -63,14 +63,18 @@ from ui.quicksearch.dlg_quicksearch_new import QuicksearchNew
 from ui.quicksearch.dlg_quicksearch_rename import QuicksearchRename
 
 # Custom modules
+from modules.api import IsogeoApiManager
 from modules.tools import Tools
+from modules.url_builder import UrlBuilder
 
 # ############################################################################
 # ########## Globals ###############
 # ##################################
 
-# utils
-tools = Tools()
+# useful submodules
+isogeo_api_mng = IsogeoApiManager()
+custom_tools = Tools()
+srv_url_bld = UrlBuilder()
 
 # LOG FILE ##
 logger = logging.getLogger()
@@ -563,7 +567,7 @@ class Isogeo:
                                  .format(result.get("query")),
                                  "Isogeo")
         # parsing
-        tags = utils.get_tags(result)
+        tags = isogeo_api_mng.get_tags(result)
         self.old_text = self.dockwidget.txt_input.text()
         # Getting the index of selected items in each combobox
         params = self.save_params()
@@ -573,7 +577,7 @@ class Isogeo:
             str(self.results_count) + self.tr(" results"))
         # Setting the number of rows in the result table
 
-        self.nb_page = str(utils.calcul_nb_page(self.results_count))
+        self.nb_page = str(custom_tools.results_pages_counter(self.results_count))
         self.dockwidget.lbl_page.setText(
             "page " + str(self.page_index) + self.tr(' on ') + self.nb_page)
 
@@ -1001,7 +1005,7 @@ class Isogeo:
         # connection is set up in QGIS
         qs = QSettings()
         if self.PostGISdict == {}:
-            self.PostGISdict = utils.build_postgis_dict(qs)
+            self.PostGISdict = srv_url_bld.build_postgis_dict(qs)
         else:
             pass
         # Looping inside the table lines. For each of them, showing the title,
@@ -1010,7 +1014,7 @@ class Isogeo:
         count = 0
         for i in result.get('results'):
             # Displaying the metadata title inside a button
-            final_text = utils.format_button_title(i.get('title'))
+            final_text = custom_tools.format_button_title(i.get('title'))
             title_button = QPushButton(final_text)
             # Connecting the button to the full metadata popup
             title_button.pressed.connect(partial(
@@ -1023,7 +1027,7 @@ class Isogeo:
             # Insert the modification date in column 2
             self.dockwidget.tbl_result.setItem(
                 count, 1, QTableWidgetItem(
-                    utils.handle_date(i.get('_modified'))))
+                    custom_tools.handle_date(i.get('_modified'))))
             # Getting the geometry
             geometry = i.get('geometry')
             if geometry is not None:
@@ -1085,7 +1089,7 @@ class Isogeo:
                 # If the data is a vector and the path is available, store
                 # useful information in the dict
                 if i.get('format') in vectorformat_list and 'path' in i:
-                    path = utils.format_path(i.get('path'))
+                    path = custom_tools.format_path(i.get('path'))
                     try:
                         open(path)
                         params = ["vector", path, i.get("title")]
@@ -1094,7 +1098,7 @@ class Isogeo:
                         pass
                 # Same if the data is a raster
                 elif i.get('format') in rasterformat_list and 'path' in i:
-                    path = utils.format_path(i.get('path'))
+                    path = custom_tools.format_path(i.get('path'))
                     try:
                         open(path)
                         params = ["raster", path]
@@ -1131,7 +1135,7 @@ class Isogeo:
                 if link.get('kind') == 'wms':
                     # Test if all the needed information is in the url.
                     url = [link.get('title'), link.get('url')]
-                    name_url = utils.build_wms_url(url)
+                    name_url = srv_url_bld.build_wms_url(url)
                     # In which case, store it in the dict.
                     if name_url != 0:
                         link_dict[u"WMS : " + name_url[1]] = name_url
@@ -1140,7 +1144,7 @@ class Isogeo:
                 # If the link is a WFS
                 elif link.get('kind') == 'wfs':
                     url = [link.get('title'), link.get('url')]
-                    name_url = utils.build_wfs_url(url)
+                    name_url = srv_url_bld.build_wfs_url(url)
                     if name_url != 0:
                         link_dict[u"WFS : " + name_url[1]] = name_url
                     else:
@@ -1152,7 +1156,7 @@ class Isogeo:
                         # WMS
                         if _link.get('kind') == 'wms':
                             url = [link.get('title'), link.get('url')]
-                            name_url = utils.build_wms_url(url)
+                            name_url = srv_url_bld.build_wms_url(url)
                             if name_url != 0:
                                 link_dict[u"WMS : " + name_url[1]] = name_url
                             else:
@@ -1160,7 +1164,7 @@ class Isogeo:
                         # WFS
                         elif _link.get('kind') == 'wfs':
                             url = [link.get('title'), link.get('url')]
-                            name_url = utils.build_wfs_url(url)
+                            name_url = srv_url_bld.build_wfs_url(url)
                             if name_url != 0:
                                 link_dict[u"WFS : " + name_url[1]] = name_url
                             else:
@@ -1191,7 +1195,7 @@ class Isogeo:
                                                       layer.get("_id")))
                                 continue
                             url = [name, path]
-                            name_url = utils.build_wfs_url(url)
+                            name_url = srv_url_bld.build_wfs_url(url)
                             if name_url != 0:
                                 link_dict[u"WFS : " + name_url[1]] = name_url
                             else:
@@ -1208,7 +1212,7 @@ class Isogeo:
                                                       layer.get("_id")))
                                 continue
                             url = [name, path]
-                            name_url = utils.build_wms_url(url)
+                            name_url = srv_url_bld.build_wms_url(url)
                             if name_url != 0:
                                 link_dict[u"WMS : " + name_url[1]] = name_url
                             else:
@@ -1236,7 +1240,7 @@ class Isogeo:
                                                       layer.get("_id")))
                                 continue
                             url = [name, path]
-                            name_url = utils.build_wfs_url(url)
+                            name_url = srv_url_bld.build_wfs_url(url)
                             if name_url != 0:
                                 link_dict[u"WFS : " + name_url[1]] = name_url
                             else:
@@ -1256,7 +1260,7 @@ class Isogeo:
                                                       layer.get("_id")))
                                 continue
                             url = [name, path]
-                            name_url = utils.build_wms_url(url)
+                            name_url = srv_url_bld.build_wms_url(url)
                             if name_url != 0:
                                 link_dict[u"WMS : " + name_url[1]] = name_url
                             else:
@@ -1589,7 +1593,7 @@ class Isogeo:
         # Info for _lang parameter
         params['lang'] = self.lang
         # URL BUILDING FUNCTION CALLED.
-        self.currentUrl = utils.build_request_url(params)
+        self.currentUrl = isogeo_api_mng.build_request_url(params)
 
         logging.info(self.currentUrl)
         # Sending the request to Isogeo API
@@ -1611,7 +1615,7 @@ class Isogeo:
                      "that is to be sent to the API")
         # Testing if the user is asking for a unexisting page (ex : page 6 out
         # of 5)
-        if self.page_index >= utils.calcul_nb_page(self.results_count):
+        if self.page_index >= custom_tools.results_pages_counter(self.results_count):
             return False
         else:
             # Adding the loading bar
@@ -1628,7 +1632,7 @@ class Isogeo:
             # Info for _lang parameter
             params['lang'] = self.lang
             # URL BUILDING FUNCTION CALLED.
-            self.currentUrl = utils.build_request_url(params)
+            self.currentUrl = isogeo_api_mng.build_request_url(params)
             # Sending the request
             if self.requestStatusClear is True:
                 self.send_request_to_isogeo_api(self.token)
@@ -1659,7 +1663,7 @@ class Isogeo:
             # Info for _lang parameter
             params['lang'] = self.lang
             # URL BUILDING FUNCTION CALLED.
-            self.currentUrl = utils.build_request_url(params)
+            self.currentUrl = isogeo_api_mng.build_request_url(params)
             # Sending the request
             if self.requestStatusClear is True:
                 self.send_request_to_isogeo_api(self.token)
@@ -1901,7 +1905,7 @@ class Isogeo:
     def show_complete_md(self, content):
         """Open the pop up window that shows the metadata sheet details."""
         logging.info("Displaying the whole metadata sheet.")
-        tags = utils.get_tags(content)
+        tags = isogeo_api_mng.get_tags(content)
         # Set the data title
         title = content.get('title')
         if title is not None:
@@ -1912,27 +1916,27 @@ class Isogeo:
         creation_date = content.get('_created')
         if creation_date is not None:
             self.IsogeoMdDetails.val_data_crea.setText(
-                utils.handle_date(creation_date))
+                custom_tools.handle_date(creation_date))
         else:
             self.IsogeoMdDetails.val_data_crea.setText('NR')
         # Set the data last modification date
         modif_date = content.get('_modified')
         if modif_date is not None:
-            self.IsogeoMdDetails.val_data_updt.setText(utils.handle_date(
+            self.IsogeoMdDetails.val_data_updt.setText(custom_tools.handle_date(
                 modif_date))
         else:
             self.IsogeoMdDetails.val_data_updt.setText('NR')
         # Set the date from which the data is valid
         valid_from = content.get('validFrom')
         if valid_from is not None:
-            self.IsogeoMdDetails.val_valid_start.setText(utils.handle_date(
+            self.IsogeoMdDetails.val_valid_start.setText(custom_tools.handle_date(
                 valid_from))
         else:
             self.IsogeoMdDetails.val_valid_start.setText('NR')
         # Set the date from which the data stops being valid
         valid_to = content.get('validTo')
         if valid_to is not None:
-            self.IsogeoMdDetails.val_valid_end.setText(utils.handle_date(
+            self.IsogeoMdDetails.val_valid_end.setText(custom_tools.handle_date(
                 valid_to))
         else:
             self.IsogeoMdDetails.val_valid_end.setText('NR')
@@ -2041,7 +2045,7 @@ class Isogeo:
         self.IsogeoMdDetails.list_events.clear()
         if content['events'] != []:
             for i in content['events']:
-                event = utils.handle_date(i['date']) + " : " + i['kind']
+                event = custom_tools.handle_date(i['date']) + " : " + i['kind']
                 if i['kind'] == 'update' and 'description' in i \
                         and i['description'] != '':
                     event += " (" + i['description'] + ")"
@@ -2172,7 +2176,7 @@ class Isogeo:
         text += u"<p>   ___________________________________________________________________   </p>"
         for share in content:
             text += u"<p><b>{0}</p></b>".format(share['name'])
-            text += self.tr(u"<p>Modified: {0}</p>").format(utils.handle_date(share['_modified']))
+            text += self.tr(u"<p>Modified: {0}</p>").format(custom_tools.handle_date(share['_modified']))
             text += self.tr(u"<p>Contact: {0}</p>").format(share['_creator']['contact']['name'])
             text += self.tr(u"<p>Applications powered by this share:</p>")
             for a in share['applications']:
@@ -2239,12 +2243,12 @@ class Isogeo:
         self.dockwidget.btn_previous.pressed.connect(self.previous_page)
         # Connecting the bug tracker button to its function
         self.dockwidget.btn_report.pressed.connect(
-            partial(utils.open_webpage,
+            partial(custom_tools.open_webpage,
                     link=self.tr(u'https://github.com/isogeo/isogeo-plugin-qgis/issues')
                     ))
 
         self.dockwidget.btn_help.pressed.connect(
-            partial(utils.open_webpage,
+            partial(custom_tools.open_webpage,
                     link="https://github.com/isogeo/isogeo-plugin-qgis/wiki"
                     ))
         # view credits
@@ -2288,12 +2292,12 @@ class Isogeo:
             partial(self.write_search_params, search_name='_default'))
 
         self.auth_prompt_form.btn_account_new.pressed.connect(partial(
-            tools.mail_to_isogeo, lang=self.lang))
+            custom_tools.mail_to_isogeo, lang=self.lang))
 
         self.dockwidget.tabWidget.currentChanged.connect(self.ask_shares_info)
 
         self.dockwidget.txt_shares.setOpenLinks(False)
-        self.dockwidget.txt_shares.anchorClicked.connect(tools.open_webpage)
+        self.dockwidget.txt_shares.anchorClicked.connect(custom_tools.open_webpage)
 
         """ --- Actions when the plugin is launched --- """
         # self.test_config_file_existence()
