+++ modules/api.py
@@ -1,0 +1,144 @@
+# -*- coding: utf-8 -*-
+
+# Standard library
+
+# PyQT
+from PyQt4.QtCore import QUrl
+
+
+class IsogeoApiManager(object):
+    """Basic class that holds utilitary methods for the plugin."""
+
+    def get_tags(self, answer):
+        """Return a tag dictionnary from the API answer.
+
+        This parse the tags contained in API_answer[tags] and class them so
+        they are more easy to handle in other function such as
+        update_fields()
+        """
+        # Initiating the dicts
+        tags = answer['tags']
+        resources_types = {}
+        owners = {}
+        keywords = {}
+        themeinspire = {}
+        formats = {}
+        srs = {}
+        actions = {}
+        # loops that sort each tag in the corresponding dict, keeping the
+        # same "key : value" structure.
+        for tag in tags.keys():
+            # owners
+            if tag.startswith('owner'):
+                owners[tag] = tags[tag]
+            # custom keywords
+            elif tag.startswith('keyword:isogeo'):
+                keywords[tag] = tags[tag]
+            # INSPIRE themes
+            elif tag.startswith('keyword:inspire-theme'):
+                themeinspire[tag] = tags[tag]
+            # formats
+            elif tag.startswith('format'):
+                formats[tag] = tags[tag]
+            # coordinate systems
+            elif tag.startswith('coordinate-system'):
+                srs[tag] = tags[tag]
+            # available actions (the last 2 are a bit specific as the value
+            # field is empty and have to be filled manually)
+            elif tag.startswith('action'):
+                if tag.startswith('action:view'):
+                    actions[tag] = u'View'
+                elif tag.startswith('action:download'):
+                    actions[tag] = u'Download'
+                elif tag.startswith('action:other'):
+                    actions[tag] = u'Other action'
+                # Test : to be removed eventually
+                else:
+                    actions[tag] = u'fonction get_tag à revoir'
+                    self.dockwidget.txt_input.setText(tag)
+            # resources type
+            elif tag.startswith('type'):
+                if tag.startswith('type:vector'):
+                    resources_types[tag] = u'Vecteur'
+                elif tag.startswith('type:raster'):
+                    resources_types[tag] = u'Raster'
+                elif tag.startswith('type:resource'):
+                    resources_types[tag] = u'Ressource'
+                elif tag.startswith('type:service'):
+                    resources_types[tag] = u'Service géographique'
+        # Creating the final object the function will return : a dictionary
+        # of dictionaries
+        new_tags = {}
+        new_tags['type'] = resources_types
+        new_tags['owner'] = owners
+        new_tags['keywords'] = keywords
+        new_tags['themeinspire'] = themeinspire
+        new_tags['formats'] = formats
+        new_tags['srs'] = srs
+        new_tags['actions'] = actions
+        return new_tags
+
+    def build_request_url(self, params):
+        """Build the request url according to the widgets."""
+        # Base url for a request to Isogeo API
+        url = 'https://v1.api.isogeo.com/resources/search?'
+        # Build the url according to the params
+        if params.get("text") != "":
+            filters = params.get("text") + " "
+        else:
+            filters = ""
+        # Owner
+        if params.get('owner') is not None:
+            filters += params.get('owner') + " "
+        # INSPIRE keywords
+        if params.get('inspire') is not None:
+            filters += params.get('inspire') + " "
+        # Format
+        if params.get('format') is not None:
+            filters += params.get('format') + " "
+        # Data type
+        if params.get('datatype') is not None:
+            filters += params.get('datatype') + " "
+        # Action : view
+        if params.get("view"):
+            filters += "action:view "
+        # Action : download
+        if params.get("download"):
+            filters += "action:download "
+        # Action : Other
+        if params.get("other"):
+            filters += "action:other "
+        # No actions
+        if params.get("noaction"):
+            filters += "has-no:action "
+        # Keywords
+        for keyword in params.get("keys"):
+            filters += keyword + " "
+        # Formating the filters
+        if filters != "":
+            filters = "q=" + filters[:-1]
+        # Geographical filter
+        if params.get("geofilter") is not None:
+            if params.get("coord") is not False:
+                filters += "&box={0}&rel={1}".format(params.get("coord"),
+                                                     params.get("operation"))
+            else:
+                pass
+        else:
+            pass
+        # Sorting order and direction
+        if params.get("show"):
+            filters += "&ob={0}&od={1}".format(params.get("ob"),
+                                               params.get("od"))
+            filters += "&_include=links,serviceLayers,layers"
+            limit = 15
+        else:
+            limit = 0
+        # Limit and offset
+        offset = (params.get("page") - 1) * 15
+        filters += "&_limit={0}&_offset={1}".format(limit, offset)
+        # Language
+        filters += "&_lang={0}".format(params.get("lang"))
+        # BUILDING FINAL URL
+        url += filters
+        return url
